apis/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
Werkzeug
PyJWT
cryptography
